Misc agda2-mode patches
The main reason for this patch is that I get a backtrace in Emacs every
time I view an Agda attachment in Gnus.

But it includes other things, such as a fix for the lisp-indent-function
setting (which applied to the buffer that happened to be current when
loading agda2-mode (e.g. when loading agda2-mode.elc), rather than to
the buffer that contains the agda2-mode.el source when editing it).

        Stefan