[ warning ] Coverage errors are non-fatal
We can recover from partial covers in the interactive mode by
adding a dummy catch-all clause with no right hand side to the
function's definition. The reduction machinery simply gets
stuck when the function is applied to arguments not matching
any of the original patterns.

Because errors appearing later on in the file may have been
caused by our being liberal, we display the non-fatal warnings
whenever a real `TypeError` is raised.